pyproject.toml
[project]
name = "pypptx-engine"
version = "0.1.0"
description = ""
authors = [
    {name = "riwrrriw",email = "awirut.p@scbx.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-pptx (>=1.0.2,<2.0.0)"
]

[project.scripts]
"pypptx-engine" = "pypptx_engine.cli:main"

[tool.poetry]
packages = [{include = "pypptx_engine", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
